+++ packages/flame/lib/src/game/base_game.dart
@@ -238,6 +238,70 @@ class BaseGame extends Game with FPSCounter {
   /// show extra information on the screen when debug mode is activated
   bool debugMode = false;
 
+  /// Changes the priority of [component] and reorders the games component list.
+  ///
+  /// Returns true if changing the component's priority modified one of the
+  /// components that existed directly on the game and false if it
+  /// either was a child of another component, if it didn't exist at all or if
+  /// it was a component added directly on the game but its priority didn't
+  /// change.
+  bool changePriority(
+    Component component,
+    int priority, {
+    bool reorderRoot = true,
+  }) {
+    if (component.priority == priority) {
+      return false;
+    }
+    if (components.contains(component)) {
+      component.changePriorityWithoutResorting(priority);
+      if (reorderRoot) {
+        components.rebalanceAll();
+      }
+      return true;
+    } else if (component is BaseComponent) {
+      component.changePriorityWithoutResorting(priority);
+      BaseComponent? foundParent;
+      for (final parent in components.whereType<BaseComponent>()) {
+        if (parent == component.parent) {
+          foundParent = parent;
+          break;
+        }
+        parent.propagateToChildren<BaseComponent>((child) {
+          if (child == component.parent) {
+            print("Found parent");
+            foundParent = child;
+            return false;
+          }
+          return true;
+        });
+        if (foundParent != null) {
+          break;
+        }
+      }
+      foundParent?.reorderChildren();
+    }
+    return false;
+  }
+
+  /// Since changing priorities is quite an expensive operation you should use
+  /// this method if you want to change multiple priorities at once so that the
+  /// tree doesn't have to be reordered multiple times.
+  void changePriorities(Map<Component, int> priorities) {
+    var hasRootComponents = false;
+    priorities.forEach((component, priority) {
+      final isRootComponent = changePriority(
+        component,
+        priority,
+        reorderRoot: false,
+      );
+      hasRootComponents |= isRootComponent;
+    });
+    if (hasRootComponents) {
+      components.rebalanceAll();
+    }
+  }
+
   /// Returns the current time in seconds with microseconds precision.
   ///
   /// This is compatible with the `dt` value used in the [update] method.